requirements.txt
python-dotenv
fastapi
httpx
uvicorn
openai
